# wasm: remove opa_number_float, introduce opa_number_mpd