This changes the hash function used for numbers -- the implication
is that set ordering of sets of numbers changes.

Also, this includes a dirty fix applied in the vendored libmpdec
sources: we're using its string formatting function now, and that
would always print the sign of the exponent. We don't want that,
so it's changed.

Fixes some of #3298, namely:

```
$ opa eval --format=pretty -t wasm 'to_number("1234567890") == 1234567890'
true
```