@srenatus 2021-04-21
fun! https://github.com/open-policy-agent/opa/runs/2400407180?check_suite_focus=true#step:3:377