@srenatus 2021-04-26
```
open-policy-agent/opa % benchstat main.bench branch.bench
name                                    old time/op  new time/op    delta
WASMArrayIteration/10-16                 408µs ± 2%     413µs ± 8%       ~     (p=0.841 n=5+5)
WASMArrayIteration/100-16                437µs ±12%     429µs ± 2%       ~     (p=0.548 n=5+5)
WASMArrayIteration/1000-16               559µs ± 5%     522µs ± 2%     -6.61%  (p=0.016 n=5+5)
WASMArrayIteration/10000-16             2.38ms ±17%    1.69ms ± 4%    -28.72%  (p=0.008 n=5+5)
WASMSetIteration/10-16                   417µs ± 4%     415µs ± 3%       ~     (p=1.000 n=5+5)
WASMSetIteration/100-16                  433µs ± 1%     431µs ± 3%       ~     (p=0.841 n=5+5)
WASMSetIteration/1000-16                 684µs ± 4%    1456µs ± 5%   +112.99%  (p=0.008 n=5+5)
WASMSetIteration/10000-16               3.69ms ± 5%  105.40ms ± 2%  +2756.01%  (p=0.008 n=5+5)
WASMObjectIteration/10-16                407µs ± 3%     428µs ± 4%     +5.13%  (p=0.016 n=5+5)
WASMObjectIteration/100-16               430µs ± 3%     447µs ± 6%     +4.08%  (p=0.016 n=5+5)
WASMObjectIteration/1000-16              668µs ± 4%    1636µs ± 2%   +144.84%  (p=0.008 n=5+5)
WASMObjectIteration/10000-16            3.94ms ± 4%  131.49ms ± 4%  +3241.28%  (p=0.008 n=5+5)
WASMLargeJSON/10x10-16                   458µs ± 5%     465µs ± 5%       ~     (p=0.690 n=5+5)
WASMLargeJSON/10x100-16                  485µs ± 4%     472µs ± 5%       ~     (p=0.222 n=5+5)
WASMLargeJSON/10x1000-16                 858µs ±14%     826µs ± 2%       ~     (p=0.421 n=5+5)
WASMLargeJSON/10x10000-16               5.12ms ± 1%    5.27ms ± 2%       ~     (p=0.063 n=4+5)
WASMLargeJSON/100x100-16                 698µs ± 7%     667µs ± 2%       ~     (p=0.421 n=5+5)
WASMLargeJSON/100x1000-16               3.17ms ± 8%    3.00ms ± 4%       ~     (p=0.222 n=5+5)
WASMVirtualDocs/total=1/hit=1-16         413µs ± 3%     420µs ± 8%       ~     (p=0.841 n=5+5)
WASMVirtualDocs/total=10/hit=1-16        415µs ± 1%     421µs ± 4%       ~     (p=0.310 n=5+5)
WASMVirtualDocs/total=100/hit=1-16       413µs ± 4%     423µs ± 4%       ~     (p=0.421 n=5+5)
WASMVirtualDocs/total=1000/hit=1-16      451µs ± 5%     473µs ± 4%       ~     (p=0.095 n=5+5)
WASMVirtualDocs/total=10/hit=10-16       414µs ± 5%     431µs ± 3%       ~     (p=0.095 n=5+5)
WASMVirtualDocs/total=100/hit=10-16      412µs ± 5%     423µs ± 5%       ~     (p=0.421 n=5+5)
WASMVirtualDocs/total=1000/hit=10-16     490µs ±11%     472µs ± 5%       ~     (p=1.000 n=5+5)
WASMVirtualDocs/total=100/hit=100-16     442µs ± 6%     435µs ± 2%       ~     (p=0.841 n=5+5)
WASMVirtualDocs/total=1000/hit=100-16    502µs ±10%     494µs ± 3%       ~     (p=0.841 n=5+5)
WASMVirtualDocs/total=1000/hit=1000-16   828µs ± 2%     807µs ± 2%       ~     (p=0.056 n=5+5)
```
We cannot merge this as-is 😞